Update uzlib and add compression support

+++ lib/uzlib/adler32.c
@@ -74,5 +74,5 @@ uint32_t uzlib_adler32(const void *data, unsigned int length, uint32_t prev_sum 
       length -= k;
    }
 
-   return (s2 << 16) | s1;
+   return ((uint32_t)s2 << 16) | s1;
 }
